# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Documentation",
        "license": {
            "name": "API"
        },
        "version": "1.0.0"
    },
    "basePath": "/public/api",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/notification/index": {
            "post": {
                "tags": [
                    "Notification"
                ],
                "summary": "Show notification",
                "description": "Show notification",
                "operationId": "Show notification",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "1 message, 2 comment, 3- order, 4- support",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/notification/markasread": {
            "post": {
                "tags": [
                    "Notification"
                ],
                "summary": "mark as read",
                "description": "mark as read",
                "operationId": "mark as read",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "notification_ides",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/notification/markasreadall": {
            "post": {
                "tags": [
                    "Notification"
                ],
                "summary": "mark as read all",
                "description": "mark as read all",
                "operationId": "mark as read all",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "1 message, 2 comment, 3- order, 4- support",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/notification/delete": {
            "post": {
                "tags": [
                    "Notification"
                ],
                "summary": "delete",
                "description": "delete",
                "operationId": "delete",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "notification_ides",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/notification/deleteall": {
            "post": {
                "tags": [
                    "Notification"
                ],
                "summary": "delete all \u0443\u0434\u0430\u043b\u0438\u0442\u044c \u0432\u0441\u0435 \u0443\u0432\u0435\u0434\u043e\u043c\u043b\u0435\u043d\u0438\u044f",
                "description": "delete all",
                "operationId": "delete all",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "1 message, 2 comment, 3- order, 4- support",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "access_token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}